// src/app/api/models/institution-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface InstitutionResponse {
  address?: string;
  code?: string;
  email?: string;
  id?: number;
  ldapBaseDn?: string;
  ldapHost?: string;
  ldapPort?: string;
  ldapUserDn?: string;
  logoLoginUrl?: string;
  logoUrl?: string;
  name?: string;
  phoneNumber?: string;
  production?: boolean;
  useLdap?: boolean;
}